fix: remove login
In a recent update (around Ronza 2024), the login process was changed to rely
on the `session.php` endpoint alone. This was determined by checking the HTTP
requests made on the browser during login.

Remove the existing `login` function and use the `requests.Session` object from
the `session.php` call, which is sufficient to provide an authenticated user
session.