refactor: remove name from error report
It was initially added because `Bot.refresh_sess` could fail, and in such
cases, we would print an error message containing `self.name`. Therefore, we
needed to declare `self.name` with an arbitrary value `None` before calling
`refresh_sess`.

However, we don't really need `self.name` inside the error message, so we
remove name, which also allows us to remove the additional `self.name`
declaration in the constructor of `Bot`.