fix: change method of fetching user data
Fetch user data using a different route, instead of the `session.php` route,
which is meant for logging in. After having logged in, calling `session.php`
again with the same session causes MouseHunt to return with an error.

Modify `Bot.refresh_sess` to log in again with a brand new session, so that we
do not repeatedly log in with an already logged in session. Modify
`Bot.get_user_data` to assume that we already have a logged in session.