fix: move captcha solver into bot project
As described in the previous commit, our requests to MouseHunt are being
blocked by Cloudflare, including the requests to fetch captcha images.
Since we already use `cloudscraper` to address this problem, move the
captcha solver into the bot project so that we can use the existing
`cloudscraper` session to fetch these images.

Moving the captcha solver into the bot has the additional benefit of
creating a monolithic application, which is better for managing and
deploying this small project. In a subsequent commit, we can clean up
the existing `captcha` directory by removing it.