chore: migrate to poetry
Use `poetry` instead of `requirements.txt` for dependency management. Update
the Dockerfile accordingly, now that we've removed `requirements.txt`. Using
`poetry` makes it easier for us to manage dev dependencies later, such as
linting tools.

While we're at it, also change all imports to be relative to the project root.
This makes it easier for modules within the Python project to import one
another, since all of the imports will be absolute imports. This will be
relevant as we refactor the project later to split the functionality into
multiple modules.