refactor: move telegram bot to client
Similar to the captcha solver, move the Telegram bot functionality into a
"client" instead, with the idea that "clients" are classes that interact with
an external service. We then inject the client in the `Bot` class for use.

Update `Settings` to handle the Telegram bot configs, as opposed to the current
method of having the `telebot.py` file call `os.environ` directly.