refactor: add game client
Add a new `GameClient` for interacting with the game via API calls. Eventually,
the `Bot` and `BotPlus` classes will be treated as services that use this
client.

To start with, the `GameClient` logs in and holds the user session and data.
`Bot` will use the `GameClient`'s session directly for now, but in future
commits we will move all of this into `GameClient`.

Note a couple quirks when parsing user data:
- When a trinket is disarmed, the value will be `None`
- Some routes return `sn_user_id` as int instead of str
- When a bait is disarmed, the value is `0` (int)