refactor: use game client for user data
Since `Bot` does not actually hold a user session anymore, we don't need
`refresh_sess`. We mostly want to refresh the session to fetch user data, so we
move all uses to `get_user_data`.

The data flow here could be simplified further, but we'll leave this for a
future refactor.