refactor: remove recursion from captcha solving
Solving captchas uses two levels of recursive calls, which makes the data flow
hard to follow. Instead of recursion, solve captchas using a simple for loop,
thus simplifying the logic greatly - at each iteration, check if a captcha
needs solving, and if so, try to do so.

This has the added benefit of allowing us to limit the number of captcha
attempts.