refactor: use game client for game ops
Move nearly all remaining interactions with the game HTTP API onto the game
client. This commit is restricted to moving the API calls only, with more
refactors coming.

We also do not change the set of journal entries functions yet, because those
will be refactored later in a significant way.