fix: always request new captcha
We only request a new captcha if `len(answer) != 5`, but we should also request
a new captcha if we weren't able to solve it the first time, since we'll still
be very likely to fail subsequent attempts. To do this, request a new captcha
before each attempt.

While we're at it, refresh the user data on each captcha request.